+++ src/main/java/com/apisql/ApiSQL/repository/SegmentoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SegmentoRepository extends JpaRepository<Segmento, Long> {
+public interface SegmentoRepository extends JpaRepository<Segmento, Integer> {
 }
